add draft logic(first version) refactor int event and note id to long

+++ src/main/java/com/gmail/netcracker/application/config/RootConfig.java
@@ -338,7 +338,7 @@ public class RootConfig {
     public RowMapper<Participant> participantRowMapper() {
         return (resultSet, i) -> {
             Participant participant = new Participant();
-            participant.setEventId(getInt(resultSet, "event_id"));
+            participant.setEventId(getLong(resultSet, "event_id"));
             participant.setPriority(getLong(resultSet, "priority_id"));
             return participant;
         };
