+++ src/main/java/com/gmail/netcracker/application/dto/dao/interfaces/EventDao.java
@@ -9,11 +9,11 @@ import java.util.List;
 public interface EventDao {
     void update(Event event);
 
-    void delete(int eventId);
+    void delete(Long eventId);
 
     void insertEvent(Event event);
 
-    Event getEvent(int eventId);
+    Event getEvent(Long eventId);
 
     List<Event> eventList();
 
@@ -27,15 +27,15 @@ public interface EventDao {
 
     List<Event> getAllMyEvents(Long personId);
 
-    void participate(Long userId, long eventId);
+    void participate(Long userId, Long eventId);
 
-    int getParticipantsCount(int eventId);
+    int getParticipantsCount(Long eventId);
 
-    List<User> getParticipants(long eventId);
+    List<User> getParticipants(Long eventId);
 
-    Participant isParticipated(Long id, int eventId);
+    Participant isParticipated(Long id, Long eventId);
 
-    void unsubscribe(long id, long eventId);
+    void unsubscribe(Long id, Long eventId);
 
     List<Event> findCreatedFriendsEvents(Long id);
 
@@ -43,13 +43,13 @@ public interface EventDao {
 
     int getMaxId();
 
-    int getEventType(int eventId);
+    int getEventType(Long eventId);
 
-    User getCreator(int eventId);
+    User getCreator(Long eventId);
 
-    Event checkCreatorById(Long personId, int eventId);
+    Event checkCreatorById(Long personId, Long eventId);
 
-    Event getEventWithPriority(Long personId, int eventId);
+    Event getEventWithPriority(Long personId, Long eventId);
 
     List<Event> listEventsWithPriority(Long personId);
